move yaml test files into their own directory
