+++ json_serializable/test/yaml_test.dart
@@ -9,11 +9,10 @@ import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
 import 'package:yaml/yaml.dart';
 
-import 'config/build_config.dart';
-
 import 'test_utils.dart';
+import 'yaml_test_files/build_config.dart';
 
-final _root = p.join('test', 'config');
+final _root = p.join('test', 'yaml_test_files');
 
 List<String> _getTests() => new Directory(_root)
     .listSync()
