+++ json_serializable/test/yaml_test_files/angular_comp.yaml
